paa-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
anthropic>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0